third_party/blink/web_tests/wpt_internal/direct-sockets/direct-sockets-enabled-on-self-origin-by-permissions-policy.https.sub.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Sockets test: can be enabled by default permissions policy (self)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/direct-sockets-helpers.js"></script>
  </head>
  <body>
    <script>
      'use strict';

      direct_sockets_test(async (t, mockDirectSocketsService) => {
        const socket = new TCPSocket('127.0.0.1', 53);
        await promise_rejects_dom(t, "NetworkError", socket.connection);
      }, "direct sockets (TCP) do not get blocked on permissions policy direct-sockets=(self)");

      direct_sockets_test(async (t, mockDirectSocketsService) => {
        const socket = new UDPSocket('127.0.0.1', 53);
        await promise_rejects_dom(t, "NetworkError", socket.connection);
      }, "direct sockets (UDP) do not get blocked on permissions policy direct-sockets=(self)");
    </script>
  </body>
</html>
